SSH connector

This connector allows to read data from SSH remote server during teal session.

Usage

ssh_connector(
  data = teal.data::teal_data(),
  join_keys = teal.data::join_keys(),
  host = NULL,
  paths = list(),
  read_expression = quote(utils::read.csv(file = path, header = TRUE)),
  title = shiny::h2(shiny::code("teal"), " - Access data from SSH")
)

Arguments

data

(teal_data) object. App developer can pass data as teal_data object with some initial data.

join_keys

(join_keys or single join_key_set) optional object with datasets column names used for joining. If empty then no joins between pairs of objects.

host

(character) an ssh server string of the form hostname[:port]. An ipv6 hostname should be wrapped in brackets like this: [2001:db8::1:80].

paths

(named list of file paths) the paths to the files containing the data on the remote server. The names of the list will define the data names that will be used to store the data on the teal.data object.

read_expression

(quoted call or expression) a call that will be used by the connector to read the contents of the file. This call can be created with quote(), substitute() or expression() functions. Note that the path variable is defined in the connector code and should be used in the expression. It defaults to utils::read.csv(file = path, header = TRUE).

When complex expression are being used, we recommend to use expression() with with each line being a separate argument. See example below.

title

(shiny.tag) a title for the connector.`

Value

object of class teal_data_module.

Examples

library(teal)
#> Loading required package: shiny
#> Loading required package: teal.data
#> Loading required package: teal.code
#> Loading required package: teal.slice
#> 
#> You are using teal version 0.15.2.9045
#> 
#> Attaching package: ‘teal’
#> The following objects are masked from ‘package:teal.slice’:
#> 
#>     as.teal_slices, teal_slices
x <- ssh_connector(
  paths = list(ADSL = "/path/to/ADSL.csv", ADTTE = "/path/to/ADTTE.csv")
)

app <- init(data = x, modules = list(example_module()))

if (interactive()) {
  shiny::runApp(app)
}

# Example with host defined explicitly
x <- ssh_connector(
  paths = list(ADSL = "/path/to/ADSL.csv", ADTTE = "/path/to/ADTTE.csv"),
  host = "localhost",
  read_expression = quote(utils::read.csv(file = path, header = TRUE))
)

app <- init(data = x, modules = list(example_module()))

if (interactive()) {
  shiny::runApp(app)
}

# Example with complex expression
x <- ssh_connector(
  paths = list(ADSL = "/path/to/ADSL.csv", ADTTE = "/path/to/ADTTE.csv"),
  read_expression = expression(
    read_custom = function(file) utils::read.csv(file = file, header = TRUE),
    read_custom(path)
  )
)

app <- init(data = x, modules = list(example_module()))

if (interactive()) {
  shiny::runApp(app)
}